app/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
crontab==0.23.0
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.19.4
pydantic==1.9.0
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
